Code is working. I still need to move a few tests, add tests for new selectors, and revise the tests for the daily value hydrograph rendering code.
